#233 Got started on model factories (BlogController)
