+++ src/Presentation/Nop.Web/Controllers/BlogController.cs
@@ -1,30 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.ServiceModel.Syndication;
 using System.Web.Mvc;
 using Nop.Core;
-using Nop.Core.Caching;
 using Nop.Core.Domain.Blogs;
 using Nop.Core.Domain.Customers;
 using Nop.Core.Domain.Localization;
-using Nop.Core.Domain.Media;
 using Nop.Services.Blogs;
-using Nop.Services.Common;
-using Nop.Services.Customers;
-using Nop.Services.Helpers;
 using Nop.Services.Localization;
 using Nop.Services.Logging;
-using Nop.Services.Media;
 using Nop.Services.Messages;
 using Nop.Services.Security;
 using Nop.Services.Seo;
 using Nop.Services.Stores;
+using Nop.Web.Factories;
 using Nop.Web.Framework;
 using Nop.Web.Framework.Controllers;
 using Nop.Web.Framework.Security;
 using Nop.Web.Framework.Security.Captcha;
-using Nop.Web.Infrastructure.Cache;
 using Nop.Web.Models.Blogs;
 
 namespace Nop.Web.Controllers
@@ -37,20 +30,16 @@ namespace Nop.Web.Controllers
         private readonly IBlogService _blogService;
         private readonly IWorkContext _workContext;
         private readonly IStoreContext _storeContext;
-        private readonly IPictureService _pictureService;
         private readonly ILocalizationService _localizationService;
-        private readonly IDateTimeHelper _dateTimeHelper;
         private readonly IWorkflowMessageService _workflowMessageService;
         private readonly IWebHelper _webHelper;
-        private readonly ICacheManager _cacheManager;
         private readonly ICustomerActivityService _customerActivityService;
         private readonly IStoreMappingService _storeMappingService;
         private readonly IPermissionService _permissionService;
+        private readonly IBlogModelFactory _blogModelFactory;
 
-        private readonly MediaSettings _mediaSettings;
         private readonly BlogSettings _blogSettings;
         private readonly LocalizationSettings _localizationSettings;
-        private readonly CustomerSettings _customerSettings;
         private readonly CaptchaSettings _captchaSettings;
 
         #endregion
@@ -60,136 +49,31 @@ namespace Nop.Web.Controllers
         public BlogController(IBlogService blogService,
             IWorkContext workContext,
             IStoreContext storeContext,
-            IPictureService pictureService,
             ILocalizationService localizationService,
-            IDateTimeHelper dateTimeHelper,
             IWorkflowMessageService workflowMessageService,
             IWebHelper webHelper,
-            ICacheManager cacheManager,
             ICustomerActivityService customerActivityService,
             IStoreMappingService storeMappingService,
             IPermissionService permissionService,
-            MediaSettings mediaSettings,
+            IBlogModelFactory blogModelFactory,
             BlogSettings blogSettings,
             LocalizationSettings localizationSettings,
-            CustomerSettings customerSettings,
             CaptchaSettings captchaSettings)
         {
             this._blogService = blogService;
             this._workContext = workContext;
             this._storeContext = storeContext;
-            this._pictureService = pictureService;
             this._localizationService = localizationService;
-            this._dateTimeHelper = dateTimeHelper;
             this._workflowMessageService = workflowMessageService;
             this._webHelper = webHelper;
-            this._cacheManager = cacheManager;
             this._customerActivityService = customerActivityService;
             this._storeMappingService = storeMappingService;
             this._permissionService = permissionService;
+            this._blogModelFactory = blogModelFactory;
 
-            this._mediaSettings = mediaSettings;
             this._blogSettings = blogSettings;
             this._localizationSettings = localizationSettings;
-            this._customerSettings = customerSettings;
             this._captchaSettings = captchaSettings;
-        }
-
-        #endregion
-
-        #region Utilities
-
-        [NonAction]
-        protected virtual void PrepareBlogPostModel(BlogPostModel model, BlogPost blogPost, bool prepareComments)
-        {
-            if (blogPost == null)
-                throw new ArgumentNullException("blogPost");
-
-            if (model == null)
-                throw new ArgumentNullException("model");
-
-            model.Id = blogPost.Id;
-            model.MetaTitle = blogPost.MetaTitle;
-            model.MetaDescription = blogPost.MetaDescription;
-            model.MetaKeywords = blogPost.MetaKeywords;
-            model.SeName = blogPost.GetSeName(blogPost.LanguageId, ensureTwoPublishedLanguages: false);
-            model.Title = blogPost.Title;
-            model.Body = blogPost.Body;
-            model.BodyOverview = blogPost.BodyOverview;
-            model.AllowComments = blogPost.AllowComments;
-            model.CreatedOn = _dateTimeHelper.ConvertToUserTime(blogPost.StartDateUtc ?? blogPost.CreatedOnUtc, DateTimeKind.Utc);
-            model.Tags = blogPost.ParseTags().ToList();
-            model.NumberOfComments = blogPost.CommentCount;
-            model.AddNewComment.DisplayCaptcha = _captchaSettings.Enabled && _captchaSettings.ShowOnBlogCommentPage;
-            if (prepareComments)
-            {
-                var blogComments = blogPost.BlogComments.OrderBy(pr => pr.CreatedOnUtc);
-                foreach (var bc in blogComments)
-                {
-                    var commentModel = new BlogCommentModel
-                    {
-                        Id = bc.Id,
-                        CustomerId = bc.CustomerId,
-                        CustomerName = bc.Customer.FormatUserName(),
-                        CommentText = bc.CommentText,
-                        CreatedOn = _dateTimeHelper.ConvertToUserTime(bc.CreatedOnUtc, DateTimeKind.Utc),
-                        AllowViewingProfiles = _customerSettings.AllowViewingProfiles && bc.Customer != null && !bc.Customer.IsGuest(),
-                    };
-                    if (_customerSettings.AllowCustomersToUploadAvatars)
-                    {
-                        commentModel.CustomerAvatarUrl = _pictureService.GetPictureUrl(
-                            bc.Customer.GetAttribute<int>(SystemCustomerAttributeNames.AvatarPictureId),
-                            _mediaSettings.AvatarPictureSize,
-                            _customerSettings.DefaultAvatarEnabled,
-                            defaultPictureType: PictureType.Avatar);
-                    }
-                    model.Comments.Add(commentModel);
-                }
-            }
-        }
-
-        [NonAction]
-        protected virtual BlogPostListModel PrepareBlogPostListModel(BlogPagingFilteringModel command)
-        {
-            if (command == null)
-                throw new ArgumentNullException("command");
-
-            var model = new BlogPostListModel();
-            model.PagingFilteringContext.Tag = command.Tag;
-            model.PagingFilteringContext.Month = command.Month;
-            model.WorkingLanguageId = _workContext.WorkingLanguage.Id;
-
-            if (command.PageSize <= 0) command.PageSize = _blogSettings.PostsPageSize;
-            if (command.PageNumber <= 0) command.PageNumber = 1;
-
-            DateTime? dateFrom = command.GetFromMonth();
-            DateTime? dateTo = command.GetToMonth();
-
-            IPagedList<BlogPost> blogPosts;
-            if (String.IsNullOrEmpty(command.Tag))
-            {
-                blogPosts = _blogService.GetAllBlogPosts(_storeContext.CurrentStore.Id,
-                    _workContext.WorkingLanguage.Id,
-                    dateFrom, dateTo, command.PageNumber - 1, command.PageSize);
-            }
-            else
-            {
-                blogPosts = _blogService.GetAllBlogPostsByTag(_storeContext.CurrentStore.Id,
-                    _workContext.WorkingLanguage.Id,
-                    command.Tag, command.PageNumber - 1, command.PageSize);
-            }
-            model.PagingFilteringContext.LoadPagedList(blogPosts);
-
-            model.BlogPosts = blogPosts
-                .Select(x =>
-                {
-                    var blogPostModel = new BlogPostModel();
-                    PrepareBlogPostModel(blogPostModel, x, false);
-                    return blogPostModel;
-                })
-                .ToList();
-
-            return model;
         }
 
         #endregion
@@ -201,7 +85,7 @@ namespace Nop.Web.Controllers
             if (!_blogSettings.Enabled)
                 return RedirectToRoute("HomePage");
 
-            var model = PrepareBlogPostListModel(command);
+            var model = _blogModelFactory.PrepareBlogPostListModel(command);
             return View("List", model);
         }
         public ActionResult BlogByTag(BlogPagingFilteringModel command)
@@ -209,7 +93,7 @@ namespace Nop.Web.Controllers
             if (!_blogSettings.Enabled)
                 return RedirectToRoute("HomePage");
 
-            var model = PrepareBlogPostListModel(command);
+            var model = _blogModelFactory.PrepareBlogPostListModel(command);
             return View("List", model);
         }
         public ActionResult BlogByMonth(BlogPagingFilteringModel command)
@@ -217,7 +101,7 @@ namespace Nop.Web.Controllers
             if (!_blogSettings.Enabled)
                 return RedirectToRoute("HomePage");
 
-            var model = PrepareBlogPostListModel(command);
+            var model = _blogModelFactory.PrepareBlogPostListModel(command);
             return View("List", model);
         }
 
@@ -258,13 +142,13 @@ namespace Nop.Web.Controllers
             //Store mapping
             if (!_storeMappingService.Authorize(blogPost))
                 return InvokeHttp404();
-
-            var model = new BlogPostModel();
-            PrepareBlogPostModel(model, blogPost, true);
-
+            
             //display "edit" (manage) link
             if (_permissionService.Authorize(StandardPermissionProvider.AccessAdminPanel) && _permissionService.Authorize(StandardPermissionProvider.ManageBlog))
                 DisplayEditLink(Url.Action("Edit", "Blog", new { id = blogPost.Id, area = "Admin" }));
+
+            var model = new BlogPostModel();
+            _blogModelFactory.PrepareBlogPostModel(model, blogPost, true);
 
             return View(model);
         }
@@ -321,7 +205,7 @@ namespace Nop.Web.Controllers
             }
 
             //If we got this far, something failed, redisplay form
-            PrepareBlogPostModel(model, blogPost, true);
+            _blogModelFactory.PrepareBlogPostModel(model, blogPost, true);
             return View(model);
         }
 
@@ -331,29 +215,8 @@ namespace Nop.Web.Controllers
             if (!_blogSettings.Enabled)
                 return Content("");
 
-            var cacheKey = string.Format(ModelCacheEventConsumer.BLOG_TAGS_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id);
-            var cachedModel = _cacheManager.Get(cacheKey, () =>
-            {
-                var model = new BlogPostTagListModel();
-
-                //get tags
-                var tags = _blogService.GetAllBlogPostTags(_storeContext.CurrentStore.Id, _workContext.WorkingLanguage.Id)
-                    .OrderByDescending(x => x.BlogPostCount)
-                    .Take(_blogSettings.NumberOfTags)
-                    .ToList();
-                //sorting
-                tags = tags.OrderBy(x => x.Name).ToList();
-
-                foreach (var tag in tags)
-                    model.Tags.Add(new BlogPostTagModel
-                    {
-                        Name = tag.Name,
-                        BlogPostCount = tag.BlogPostCount
-                    });
-                return model;
-            });
-
-            return PartialView(cachedModel);
+            var model = _blogModelFactory.PrepareBlogPostTagListModel();
+            return PartialView(model);
         }
 
         [ChildActionOnly]
@@ -362,61 +225,8 @@ namespace Nop.Web.Controllers
             if (!_blogSettings.Enabled)
                 return Content("");
 
-            var cacheKey = string.Format(ModelCacheEventConsumer.BLOG_MONTHS_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id);
-            var cachedModel = _cacheManager.Get(cacheKey, () =>
-            {
-                var model = new List<BlogPostYearModel>();
-
-                var blogPosts = _blogService.GetAllBlogPosts(_storeContext.CurrentStore.Id,
-                    _workContext.WorkingLanguage.Id);
-                if (blogPosts.Any())
-                {
-                    var months = new SortedDictionary<DateTime, int>();
-
-                    var blogPost = blogPosts[blogPosts.Count - 1];
-                    var first = blogPost.StartDateUtc ?? blogPost.CreatedOnUtc;
-                    while (DateTime.SpecifyKind(first, DateTimeKind.Utc) <= DateTime.UtcNow.AddMonths(1))
-                    {
-                        var list = blogPosts.GetPostsByDate(new DateTime(first.Year, first.Month, 1), new DateTime(first.Year, first.Month, 1).AddMonths(1).AddSeconds(-1));
-                        if (list.Any())
-                        {
-                            var date = new DateTime(first.Year, first.Month, 1);
-                            months.Add(date, list.Count);
-                        }
-
-                        first = first.AddMonths(1);
-                    }
-
-
-                    int current = 0;
-                    foreach (var kvp in months)
-                    {
-                        var date = kvp.Key;
-                        var blogPostCount = kvp.Value;
-                        if (current == 0)
-                            current = date.Year;
-
-                        if (date.Year > current || !model.Any())
-                        {
-                            var yearModel = new BlogPostYearModel
-                            {
-                                Year = date.Year
-                            };
-                            model.Insert(0, yearModel);
-                        }
-
-                        model.First().Months.Insert(0, new BlogPostMonthModel
-                        {
-                            Month = date.Month,
-                            BlogPostCount = blogPostCount
-                        });
-
-                        current = date.Year;
-                    }
-                }
-                return model;
-            });
-            return PartialView(cachedModel);
+            var model = _blogModelFactory.PrepareBlogPostYearModel();
+            return PartialView(model);
         }
 
         [ChildActionOnly]
