+++ src/Presentation/Nop.Web/Factories/BlogModelFactory.cs
@@ -1,0 +1,247 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Nop.Core;
+using Nop.Core.Caching;
+using Nop.Core.Domain.Blogs;
+using Nop.Core.Domain.Customers;
+using Nop.Core.Domain.Media;
+using Nop.Services.Blogs;
+using Nop.Services.Common;
+using Nop.Services.Customers;
+using Nop.Services.Helpers;
+using Nop.Services.Media;
+using Nop.Services.Seo;
+using Nop.Web.Framework.Security.Captcha;
+using Nop.Web.Infrastructure.Cache;
+using Nop.Web.Models.Blogs;
+
+namespace Nop.Web.Factories
+{
+    public partial class BlogModelFactory : IBlogModelFactory
+    {
+        #region Fields
+
+        private readonly IBlogService _blogService;
+        private readonly IWorkContext _workContext;
+        private readonly IStoreContext _storeContext;
+        private readonly IPictureService _pictureService;
+        private readonly IDateTimeHelper _dateTimeHelper;
+        private readonly ICacheManager _cacheManager;
+
+        private readonly MediaSettings _mediaSettings;
+        private readonly BlogSettings _blogSettings;
+        private readonly CustomerSettings _customerSettings;
+        private readonly CaptchaSettings _captchaSettings;
+
+        #endregion
+
+        #region Constructors
+
+        public BlogModelFactory(IBlogService blogService,
+            IWorkContext workContext,
+            IStoreContext storeContext,
+            IPictureService pictureService,
+            IDateTimeHelper dateTimeHelper,
+            ICacheManager cacheManager,
+            MediaSettings mediaSettings,
+            BlogSettings blogSettings,
+            CustomerSettings customerSettings,
+            CaptchaSettings captchaSettings)
+        {
+            this._blogService = blogService;
+            this._workContext = workContext;
+            this._storeContext = storeContext;
+            this._pictureService = pictureService;
+            this._dateTimeHelper = dateTimeHelper;
+            this._cacheManager = cacheManager;
+            this._mediaSettings = mediaSettings;
+            this._blogSettings = blogSettings;
+            this._customerSettings = customerSettings;
+            this._captchaSettings = captchaSettings;
+        }
+
+        #endregion
+
+        #region Methods
+
+        public virtual void PrepareBlogPostModel(BlogPostModel model, BlogPost blogPost, bool prepareComments)
+        {
+            if (model == null)
+                throw new ArgumentNullException("model");
+
+            if (blogPost == null)
+                throw new ArgumentNullException("blogPost");
+
+            model.Id = blogPost.Id;
+            model.MetaTitle = blogPost.MetaTitle;
+            model.MetaDescription = blogPost.MetaDescription;
+            model.MetaKeywords = blogPost.MetaKeywords;
+            model.SeName = blogPost.GetSeName(blogPost.LanguageId, ensureTwoPublishedLanguages: false);
+            model.Title = blogPost.Title;
+            model.Body = blogPost.Body;
+            model.BodyOverview = blogPost.BodyOverview;
+            model.AllowComments = blogPost.AllowComments;
+            model.CreatedOn = _dateTimeHelper.ConvertToUserTime(blogPost.StartDateUtc ?? blogPost.CreatedOnUtc, DateTimeKind.Utc);
+            model.Tags = blogPost.ParseTags().ToList();
+            model.NumberOfComments = blogPost.CommentCount;
+            model.AddNewComment.DisplayCaptcha = _captchaSettings.Enabled && _captchaSettings.ShowOnBlogCommentPage;
+            if (prepareComments)
+            {
+                var blogComments = blogPost.BlogComments.OrderBy(pr => pr.CreatedOnUtc);
+                foreach (var bc in blogComments)
+                {
+                    var commentModel = new BlogCommentModel
+                    {
+                        Id = bc.Id,
+                        CustomerId = bc.CustomerId,
+                        CustomerName = bc.Customer.FormatUserName(),
+                        CommentText = bc.CommentText,
+                        CreatedOn = _dateTimeHelper.ConvertToUserTime(bc.CreatedOnUtc, DateTimeKind.Utc),
+                        AllowViewingProfiles = _customerSettings.AllowViewingProfiles && bc.Customer != null && !bc.Customer.IsGuest(),
+                    };
+                    if (_customerSettings.AllowCustomersToUploadAvatars)
+                    {
+                        commentModel.CustomerAvatarUrl = _pictureService.GetPictureUrl(
+                            bc.Customer.GetAttribute<int>(SystemCustomerAttributeNames.AvatarPictureId),
+                            _mediaSettings.AvatarPictureSize,
+                            _customerSettings.DefaultAvatarEnabled,
+                            defaultPictureType: PictureType.Avatar);
+                    }
+                    model.Comments.Add(commentModel);
+                }
+            }
+        }
+
+        public virtual BlogPostListModel PrepareBlogPostListModel(BlogPagingFilteringModel command)
+        {
+            if (command == null)
+                throw new ArgumentNullException("command");
+
+            var model = new BlogPostListModel();
+            model.PagingFilteringContext.Tag = command.Tag;
+            model.PagingFilteringContext.Month = command.Month;
+            model.WorkingLanguageId = _workContext.WorkingLanguage.Id;
+
+            if (command.PageSize <= 0) command.PageSize = _blogSettings.PostsPageSize;
+            if (command.PageNumber <= 0) command.PageNumber = 1;
+
+            DateTime? dateFrom = command.GetFromMonth();
+            DateTime? dateTo = command.GetToMonth();
+
+            IPagedList<BlogPost> blogPosts;
+            if (String.IsNullOrEmpty(command.Tag))
+            {
+                blogPosts = _blogService.GetAllBlogPosts(_storeContext.CurrentStore.Id,
+                    _workContext.WorkingLanguage.Id,
+                    dateFrom, dateTo, command.PageNumber - 1, command.PageSize);
+            }
+            else
+            {
+                blogPosts = _blogService.GetAllBlogPostsByTag(_storeContext.CurrentStore.Id,
+                    _workContext.WorkingLanguage.Id,
+                    command.Tag, command.PageNumber - 1, command.PageSize);
+            }
+            model.PagingFilteringContext.LoadPagedList(blogPosts);
+
+            model.BlogPosts = blogPosts
+                .Select(x =>
+                {
+                    var blogPostModel = new BlogPostModel();
+                    PrepareBlogPostModel(blogPostModel, x, false);
+                    return blogPostModel;
+                })
+                .ToList();
+
+            return model;
+        }
+
+        public virtual BlogPostTagListModel PrepareBlogPostTagListModel()
+        {
+            var cacheKey = string.Format(ModelCacheEventConsumer.BLOG_TAGS_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id);
+            var cachedModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var model = new BlogPostTagListModel();
+
+                //get tags
+                var tags = _blogService.GetAllBlogPostTags(_storeContext.CurrentStore.Id, _workContext.WorkingLanguage.Id)
+                    .OrderByDescending(x => x.BlogPostCount)
+                    .Take(_blogSettings.NumberOfTags)
+                    .ToList();
+                //sorting
+                tags = tags.OrderBy(x => x.Name).ToList();
+
+                foreach (var tag in tags)
+                    model.Tags.Add(new BlogPostTagModel
+                    {
+                        Name = tag.Name,
+                        BlogPostCount = tag.BlogPostCount
+                    });
+                return model;
+            });
+
+            return cachedModel;
+        }
+
+        public virtual List<BlogPostYearModel> PrepareBlogPostYearModel()
+        {
+            var cacheKey = string.Format(ModelCacheEventConsumer.BLOG_MONTHS_MODEL_KEY, _workContext.WorkingLanguage.Id, _storeContext.CurrentStore.Id);
+            var cachedModel = _cacheManager.Get(cacheKey, () =>
+            {
+                var model = new List<BlogPostYearModel>();
+
+                var blogPosts = _blogService.GetAllBlogPosts(_storeContext.CurrentStore.Id,
+                    _workContext.WorkingLanguage.Id);
+                if (blogPosts.Any())
+                {
+                    var months = new SortedDictionary<DateTime, int>();
+
+                    var blogPost = blogPosts[blogPosts.Count - 1];
+                    var first = blogPost.StartDateUtc ?? blogPost.CreatedOnUtc;
+                    while (DateTime.SpecifyKind(first, DateTimeKind.Utc) <= DateTime.UtcNow.AddMonths(1))
+                    {
+                        var list = blogPosts.GetPostsByDate(new DateTime(first.Year, first.Month, 1), new DateTime(first.Year, first.Month, 1).AddMonths(1).AddSeconds(-1));
+                        if (list.Any())
+                        {
+                            var date = new DateTime(first.Year, first.Month, 1);
+                            months.Add(date, list.Count);
+                        }
+
+                        first = first.AddMonths(1);
+                    }
+
+
+                    int current = 0;
+                    foreach (var kvp in months)
+                    {
+                        var date = kvp.Key;
+                        var blogPostCount = kvp.Value;
+                        if (current == 0)
+                            current = date.Year;
+
+                        if (date.Year > current || !model.Any())
+                        {
+                            var yearModel = new BlogPostYearModel
+                            {
+                                Year = date.Year
+                            };
+                            model.Insert(0, yearModel);
+                        }
+
+                        model.First().Months.Insert(0, new BlogPostMonthModel
+                        {
+                            Month = date.Month,
+                            BlogPostCount = blogPostCount
+                        });
+
+                        current = date.Year;
+                    }
+                }
+                return model;
+            });
+            return cachedModel;
+        }
+
+        #endregion
+    }
+}
