+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -269,6 +269,8 @@
     <Compile Include="Extensions\HtmlExtensions.cs" />
     <Compile Include="Extensions\ControllerExtensions.cs" />
     <Compile Include="Extensions\MappingExtensions.cs" />
+    <Compile Include="Factories\IBlogModelFactory.cs" />
+    <Compile Include="Factories\BlogModelFactory.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
